Use separate eventFilter to handle application-wise events.
Added a mechanism to lock the window drag object, when one candidate event is recieved, so that 
when this event is propagated to parent widgets, it is not checked again. Lock is removed at 
next mouseReleaseEvent

Never eat mousePress events, since they are not used (directly) to trigger the grab.

Cleaned up blackListing using the new functionalities above.

Fixed a number of conflicts between apps' events and window grabbing, using the functionalities 
above.

CCBUG: 21415
 --This line, and those below, will be ignored--

M    oxygenwindowmanager.h
M    oxygenwindowmanager.cpp

svn path=/trunk/KDE/kdebase/workspace/; revision=1137112