Fix crash on exit in all QCommandLineParser-based programs.
Example: kioclient5 ls  (= a syntax error).

The issue is that ::exit(1) is called, so ~QGuiApplication isn't called
(so the style isn't deleted), but global static objects are deleted,
which deletes the style-plugin-factory, which unloads the style plugin.

The crash happened because "AppEventFilter" in oxygen would still be
installed as an app event filter, but the plugin was unloaded, so any
calls to it would crash.