Add option to enable high DPI pixmaps in oxygen-demo
This is the flag that application developers need to set if they want to
support smooth scaling.

From a QStyle perspective we need to support both having this set and
not set.

With QT_DEVICE_PIXEL_RATIO=2:
Without this flag everything looks blocky but is in the right place.

With this flag, everything looks super smooth but some layout parts are
screwed.

REVIEW: 120234