Remove outdated hack deleting the currently used QStyle at teardown
According to the comment we need to delete the qstyle object before the
plugin if the application exits with exit(). A small test showed this is
no longer the case.

More importantly this fixes a crash we see when styles were switched at
runtime and the app uses exit().

Because Breeze has a similar delete line the first would run deleting
the style object but leaving a dangling pointer, the second would run
and crash.