Create shadows unconditionally
Oxygen checks if shadows are supported by checking for a property on the
root window. This is set by kwin at some point after startup.

This leaves a gap for apps/plasmashell to start, determine shadows are
not supported and then never render them for the lifespan of that app.

It also can't handle window managers being replaced.

Simply not checking solves all the issues with limited drawback

REVIEW: 123284