[kstyle] Disable window move on Wayland
Unfortunately we cannot support moving windows from the style on Wayland
(yet). To move a window we need the (Xdg)ShellSurface and the wl_seat.
Through the integration plugin we can get the ShellSurface, but not the
seat. Thus we cannot invoke the feature through KWayland.

We either need to extend the native interface and/or extend QWindow to
expose a startSystemMove (QPlatformWindow already supports a
startSystemResize).