oxygen windeco: Fix use-after-free for pixmap in QCache.
Coverity identified (in CID 1336171) a usage of QPixmap after it's been
the parameter to QCache<QPixmap>::insert.

This is unsafe since QCache gains ownership of whatever object is passed
in, and can even (at least in theory) delete the object you pass in.
Instead give the QCache its own copy of the pixmap.

To make this work without leaking our existing pixmap we need to work
with a QPixmap instead of QPixmap* in the non-cached case. This leads to
whitespace changes; use git diff -b to filter pure whitespace changes
out.

REVIEW:127837