Fix crash on exit
Since Qt 5.6.0, Qt5 applications started crashing on exit. All signs point to this delete-on-destroy hack which was added to avoid outliving the plugin lifetime.

This method is wrong because the returned style is owned by the caller (QApplication, QProxyStyle, etc) and will cleaned up when those users are destructed.

REVIEW: 128760
CCMAIL: peter@lekensteyn.nl