[kstyle] Implement window moving on Wayland
Summary:
So far oxygen hard disabled the window moving on Wayland. With this
change the required functionality gets added.

For that Oxygen creates an additional Seat and a Pointer on it to track
all pointer button events on the window. That way the kstyle gets the
latest serial which needs to be passed to the move requests. This is not
available through QtWayland's native interface, thus Oxygen needs to
interact with Wayland directly.

When the move is triggered Oxygen gets the ShellSurface for the window
and triggers the move on the own Seat object with the tracked serial.

Test Plan: Tested with KWin/Wayland

Reviewers: #plasma, hpereiradacosta

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3096