Fix indentation of icon for left aligned toolbuttons
Summary:
Gwenview's {nav Operations} sidebar uses custom `QToolButtons` which are
wide and left aligned. Special support for this in Oxygen was added with
b33942690c65. However, compared to the KDE4 version there is not enough space
between the left button border and the icon (Screenshots from the
matching Diff for Breeze, see D9281):

{F5539728}

This is fixed by adding the appropriate margin:

{F5539730}

Note how this replicates exactly the spacing of the toolbuttons in the
regular toolbar on the top (which had its icon size set to "small" in
this screenshot).

CCBUG: 381535

Test Plan:
Icons on toolbuttons in Gwenview's {nav Operations} sidebar no longer
squished all the way to the left, text spacing still looking good.

Reviewers: hpereiradacosta

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9282