Don't drag windows in empty areas from touch/pen events
For a mouse users have the precision that if they clicked in an empty
area they want to perform some action.

For touch and tablets that isn't necessarily true.

From Boud in a kwin report: "The drag the window by empty areas is a
nasty one as well, especially when you're using a pen."