[kstyle] Drop Qt 4 style plugin
Summary:
KWindowSystem cannot be used in the Qt 4 style plugin. On the other hand,
we need to use KWindowSystem to make QWidget-based internal clients in
KWin cast drop-shadows.

Another problem with the Qt 4 style plugin is that some distributions
have already dropped Qt 4 (it's been EOL for almost 5 years), so one has
to build it in order to verify that his or her change works with Qt 4.

Given maintenance burden and the fact that Qt 6 is around the corner,
this change drops the Qt 4 style plugin.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Maniphest Tasks: T12496

Differential Revision: https://phabricator.kde.org/D26621