Fix cursor left_ptr cutting off under extreme magnification
[#69](https://invent.kde.org/plasma/oxygen/-/merge_requests/69) introduces SVG cursors, which makes small imperfections of the oxygen cursor noticeable.

Specifically the "Shake cursor to find" feature shows the top and bottom of the left_ptr cursor having obvious lines where the rounded tip and shadow are cut off, with less obvious cut-off lines to the left and right in the shadow.

![image](/uploads/19e1dc3a742dad6b5a41b528ea34ed42/image.png){width=300 height=300}

This MR adds a few minimal changes to the SVG of left_ptr:

- expanding the svg canvas area from 24x24 to 32x32
- Moving all path definition and other positioned elements down by 5 px
- removing a clip-plane from the shadow that is not needed with the expanded canvas
- increasing the area of the gaussian filter applied to the shadow (which also caused clipping of the shadow)

The left_ptr.in is modified to have its hotspot at 5,5 instead of 5,0.

![image](/uploads/da51cb38ae2ecb7f26ff0fe10d3d3e84/image.png){width=300 height=300} ![image](/uploads/bd63c16d2c382fa5396a167427ce93c8/image.png){width=300 height=300}

Above the new cursor, and a filter applied to confirm there is no clipping with the border.

I have (now) verified using layers in GIMP that the old and new cursors are pixel-perfectly identical under 50x magnification as shown here, apart from the clipping and movement.